refactor: 컨트롤러 네이밍 변경, 패키지 경로 변경
