+++ src/main/java/woowacourse/shoppingcart/ui/OrderController.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.documentation;
+package woowacourse.shoppingcart.ui;
 
 import java.net.URI;
 import java.util.List;
