+++ src/main/java/woowacourse/shoppingcart/ui/ProductController.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.documentation;
+package woowacourse.shoppingcart.ui;
 
 import java.net.URI;
 import org.springframework.http.HttpStatus;
